Also moved ComputerPickle and FilePathPickle to workflow-durable-task-step, introducing FilePathUtils pending JENKINS-26096. Originally-Committed-As: 712d524bf0942c0d7068c829a3a9ca45b7dae329
